requirements.txt
fastapi
uvicorn
neo4j
sentence-transformers
openai
pandas
python-dotenv